banking-agentic-rag/requirements.txt
langchain
langchain-core
langchain-text-splitters
langgraph
langsmith
openai
pydantic
pydantic_core
python-dotenv
SQLAlchemy
uvicorn
faiss-cpu
rank_bm25
tf-keras